Create reusable lib to create endoint getter
